use-map-legends in fishing-map
