+++ packages/react-hooks/src/use-map-legend/index.ts
@@ -1,7 +1,1 @@
-import { LayerMetadataLegend } from '@globalfishingwatch/layer-composer'
-
 export { default } from './use-map-legend'
-
-export type LegendLayer = LayerMetadataLegend & {
-  color: string
-}
