+++ packages/react-hooks/src/use-map-legend/use-map-legend.ts
@@ -1,39 +1,77 @@
 import { useMemo } from 'react'
-import { ExtendedStyle } from '@globalfishingwatch/layer-composer'
+import { ExtendedStyle, Generators, LayerMetadataLegend } from '@globalfishingwatch/layer-composer'
+import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
+import type {
+  LegendLayer,
+  LegendLayerBivariate,
+} from '@globalfishingwatch/ui-components/dist/map-legend'
 import { InteractionEvent } from '../use-map-interaction'
-import { LegendLayer } from '.'
 
-export function getLegendLayers(
-  style: ExtendedStyle,
-  event?: InteractionEvent
-): LegendLayer[] | undefined {
-  const layers = style?.layers?.flatMap((layer) => {
+const getLegendLayers = (
+  style?: ExtendedStyle,
+  dataviews?: UrlDataviewInstance[],
+  hoveredEvent?: InteractionEvent | null
+) => {
+  if (!style) return []
+  return style.layers?.flatMap((layer) => {
     if (!layer.metadata?.legend) return []
-    const legendLayer = {
-      ...layer.metadata.legend,
-      color: layer.metadata.color || 'red',
-      generatorId: layer.metadata.generatorId as string,
-    }
-    if (!event) return legendLayer
-    const eventFeature = event?.features?.find(
-      (feature) => feature.generatorId === layer.metadata?.generatorId
-    )
-    if (!eventFeature) return legendLayer
-    return {
-      ...legendLayer,
-      currentValue: eventFeature.value,
-      properties: eventFeature.properties,
-    }
+
+    const sublayerLegendsMetadata: LayerMetadataLegend[] = Array.isArray(layer.metadata.legend)
+      ? layer.metadata.legend
+      : [layer.metadata.legend]
+    return sublayerLegendsMetadata.map((sublayerLegendMetadata, sublayerIndex) => {
+      const id = sublayerLegendMetadata.id || (layer.metadata?.generatorId as string)
+      const dataview = dataviews?.find((d) => d.id === id)
+
+      const sublayerLegend: LegendLayer | LegendLayerBivariate = {
+        ...sublayerLegendMetadata,
+        id: `legend_${id}`,
+        color: layer.metadata?.color || dataview?.config?.color || 'red',
+      }
+
+      const generatorType = layer.metadata?.generatorType
+
+      if (generatorType === Generators.Type.Heatmap) {
+        const value = hoveredEvent?.features?.find(
+          (f) => f.generatorId === layer.metadata?.generatorId
+        )?.value
+        if (value) {
+          sublayerLegend.currentValue = value
+        }
+      } else if (generatorType === Generators.Type.HeatmapAnimated) {
+        const getHoveredFeatureValueForSublayerIndex = (index: number): number => {
+          const hoveredFeature = hoveredEvent?.features?.find(
+            (f) => f.temporalgrid?.sublayerIndex === index
+          )
+          return hoveredFeature?.value
+        }
+        // Both bivariate sublayers come in the same sublayerLegend (see getLegendsBivariate in LC)
+        if (sublayerLegend.type === 'bivariate') {
+          sublayerLegend.currentValues = [
+            getHoveredFeatureValueForSublayerIndex(0),
+            getHoveredFeatureValueForSublayerIndex(1),
+          ]
+        } else {
+          sublayerLegend.currentValue = getHoveredFeatureValueForSublayerIndex(sublayerIndex)
+        }
+      } else if (generatorType === Generators.Type.UserContext) {
+        // TODO use dataset propertyToInclude value
+        sublayerLegend.label = ''
+      }
+      return sublayerLegend
+    })
   })
-  return layers as LegendLayer[]
 }
 
-function useMapLegend(style?: ExtendedStyle, interactionEvent?: InteractionEvent) {
+function useMapLegend(
+  style?: ExtendedStyle,
+  dataviews?: UrlDataviewInstance[],
+  interactionEvent?: InteractionEvent | null
+) {
   const legendLayers = useMemo(() => {
     if (!style) return
-    const legendLayers = getLegendLayers(style, interactionEvent)
-    return legendLayers?.reverse()
-  }, [style, interactionEvent])
+    return getLegendLayers(style, dataviews, interactionEvent)
+  }, [style, dataviews, interactionEvent])
   return legendLayers
 }
 export default useMapLegend
